+++ src/dashboardWebView/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useCard';
 export * from './useExtensibility';
 export * from './useMedia';
 export * from './useMessages';