+++ src/helpers/FilesHelper.ts
@@ -1,6 +1,6 @@
 import { Notifications } from './Notifications';
 import { Uri } from 'vscode';
-import { Folders } from '../commands/Folders';
+import { Folders, WORKSPACE_PLACEHOLDER } from '../commands/Folders';
 import { isValidFile } from './isValidFile';
 import { parseWinPath } from './parseWinPath';
 import { join } from 'path';
@@ -40,5 +40,16 @@ export class FilesHelper {
     const wsFolder = Folders.getWorkspaceFolder();
     let absPath = join(parseWinPath(wsFolder?.fsPath || ''), filePath);
     return parseWinPath(absPath);
+  }
+
+  /**
+   * Absolute path to rel path
+   * @param filePath
+   * @returns
+   */
+  public static absToRelPath(filePath: string): string {
+    const wsFolder = Folders.getWorkspaceFolder();
+    const relPath = filePath.replace(wsFolder?.fsPath || '', WORKSPACE_PLACEHOLDER);
+    return parseWinPath(relPath);
   }
 }