+++ src/services/index.ts
@@ -1,5 +1,6 @@
 export * from './Credentials';
 export * from './ModeSwitch';
 export * from './PagesParser';
+export * from './PinnedItems';
 export * from './SponsorAI';
 export * from './Terminal';